chore: Delete analysis, data, and metric.
With everything split out into plugins, we can finally delete almost all
of the `analysis`, `data`, and `metric` modules. A little of Git and NPM
interaction code has been preserved under `util/`, and scoring code has been
moved to its own top-level module (it had previously been under `analysis`).
