--- hipcheck/src/metric/fuzz.rs
@@ -1,26 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-use crate::{
-	data::Fuzz,
-	error::{Context as _, Result},
-	metric::MetricProvider,
-};
-use serde::Serialize;
-use std::sync::Arc;
-
-#[derive(Debug, Eq, PartialEq, Serialize)]
-pub struct FuzzOutput {
-	pub fuzz_result: Fuzz,
-}
-
-pub fn fuzz_metric(db: &dyn MetricProvider) -> Result<Arc<FuzzOutput>> {
-	log::debug!("running fuzz metric");
-
-	let fuzz_result = db
-		.fuzz_check()
-		.context("failed to get response from fuzz check")?;
-
-	log::info!("completed fuzz metric");
-
-	Ok(Arc::new(FuzzOutput { fuzz_result }))
-}
