+++ hipcheck/src/util/mod.rs
@@ -4,8 +4,9 @@
 
 pub mod command;
 pub mod fs;
+pub mod git;
 pub mod http;
 pub mod kdl;
-pub mod redacted;
+pub mod npm;
 #[cfg(test)]
 pub mod test;
